+++ apis/enigma-mocker/src/index.js
@@ -1,0 +1,5 @@
+import fromGenericObjects from './from-generic-objects';
+
+export default {
+  fromGenericObjects,
+};