+++ apis/enigma-mocker/src/util.js
@@ -1,0 +1,7 @@
+/* eslint-disable import/prefer-default-export */
+export async function findAsync(arr, asyncCallback) {
+  const promises = arr.map(asyncCallback);
+  const results = await Promise.all(promises);
+  const index = results.findIndex((result) => result);
+  return arr[index];
+}