+++ apis/nucleus/src/components/Cell.jsx
@@ -15,6 +15,7 @@ import useRect from '../hooks/useRect';
 import useLayout, { useAppLayout } from '../hooks/useLayout';
 import InstanceContext from '../contexts/InstanceContext';
 import useObjectSelections from '../hooks/useObjectSelections';
+import eventmixin from '../selections/event-mixin';
 
 /**
  * @interface
@@ -92,7 +93,7 @@ const LoadingSn = ({ delay = 750 }) => {
     return () => clearTimeout(handle);
   }, []);
 
-  return showLoading && <Loading />;
+  return showLoading ? <Loading /> : null;
 };
 
 const handleModal = ({ sn, layout, model }) => {
@@ -254,7 +255,7 @@ const getType = async ({ types, name, version }) => {
   return SN;
 };
 
-const loadType = async ({ dispatch, types, visualization, version, model, app, selections, blurCallback, nebbie }) => {
+const loadType = async ({ dispatch, types, visualization, version, model, app, selections, nebbie, focusHandler }) => {
   try {
     const snType = await getType({ types, name: visualization, version });
     const sn = snType.create({
@@ -262,7 +263,7 @@ const loadType = async ({ dispatch, types, visualization, version, model, app, s
       app,
       selections,
       nebbie,
-      blurCallback,
+      focusHandler,
     });
     return sn;
   } catch (err) {
@@ -282,12 +283,32 @@ const Cell = forwardRef(
     const [state, dispatch] = useReducer(contentReducer, initialState(initialError));
     const [layout, { validating, canCancel, canRetry }, longrunning] = useLayout(model);
     const [appLayout] = useAppLayout(app);
-    const [contentRef, contentRect] = useRect();
+    const [contentRef, contentRect, contentNode] = useRect();
     const [snOptions, setSnOptions] = useState(initialSnOptions);
     const [snPlugins, setSnPlugins] = useState(initialSnPlugins);
     const [selections] = useObjectSelections(app, model);
     const [hovering, setHover] = useState(false);
     const hoveringDebouncer = useRef({ enter: null, leave: null });
+    const focusHandler = useRef({
+      focusToolbarButton(last) {
+        // eslint-disable-next-line react/no-this-in-sfc
+        this.emit(last ? 'focus_toolbar_last' : 'focus_toolbar_first');
+      },
+    });
+
+    useEffect(() => {
+      eventmixin(focusHandler.current);
+    }, []);
+
+    focusHandler.current.blurCallback = (resetFocus) => {
+      halo.root.toggleFocusOfCells();
+      if (resetFocus && contentNode) {
+        contentNode.focus();
+      }
+    };
+    focusHandler.current.refocusContent = () => {
+      state.sn.component && typeof state.sn.component.focus === 'function' && state.sn.component.focus();
+    };
 
     const handleOnMouseEnter = () => {
       if (hoveringDebouncer.current.leave) {
@@ -317,13 +338,6 @@ const Cell = forwardRef(
       }
     };
 
-    const relinquishFocus = (resetFocus) => {
-      halo.root.toggleFocusOfCells();
-      if (resetFocus && cellNode) {
-        cellNode.focus();
-      }
-    };
-
     useEffect(() => {
       if (initialError || !appLayout || !layout) {
         return undefined;
@@ -348,7 +362,7 @@ const Cell = forwardRef(
           app,
           selections,
           nebbie,
-          blurCallback: relinquishFocus,
+          focusHandler: focusHandler.current,
         });
         if (sn) {
           dispatch({ type: 'LOADED', sn, visualization });
@@ -464,14 +478,12 @@ const Cell = forwardRef(
     return (
       <Paper
         style={{ position: 'relative', width: '100%', height: '100%', overflow: 'hidden' }}
-        tabIndex={keyboardNavigation ? 0 : -1}
         elevation={0}
         square
         className={CellElement.className}
         ref={cellRef}
         onMouseEnter={handleOnMouseEnter}
         onMouseLeave={handleOnMouseLeave}
-        onKeyDown={keyboardNavigation ? handleKeyDown : null}
       >
         <Grid
           container
@@ -486,11 +498,19 @@ const Cell = forwardRef(
           }}
         >
           {cellNode && layout && state.sn && (
-            <Header layout={layout} sn={state.sn} anchorEl={cellNode} hovering={hovering}>
+            <Header
+              layout={layout}
+              sn={state.sn}
+              anchorEl={cellNode}
+              hovering={hovering}
+              focusHandler={focusHandler.current}
+            >
               &nbsp;
             </Header>
           )}
           <Grid
+            tabIndex={keyboardNavigation ? 0 : -1}
+            onKeyDown={keyboardNavigation ? handleKeyDown : null}
             item
             xs
             style={{