+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.2.0](https://github.com/qlik-oss/nebula.js/compare/v2.1.0...v2.2.0) (2021-11-17)
+
+### Features
+
+- **engima-mock:** mock qlik engine ([#705](https://github.com/qlik-oss/nebula.js/issues/705)) ([bba840c](https://github.com/qlik-oss/nebula.js/commit/bba840c80f269abf3580cb261dfb67c3e36c3bac))
+- focus in and out of selection toolbar ([#701](https://github.com/qlik-oss/nebula.js/issues/701)) ([9b2a038](https://github.com/qlik-oss/nebula.js/commit/9b2a03804f35ad60303362a47fca217ba232de5b))
+
 # [2.1.0](https://github.com/qlik-oss/nebula.js/compare/v2.0.0...v2.1.0) (2021-10-20)
 
 ### Bug Fixes